+++ src/data/rule/buildRule.js
@@ -3,12 +3,12 @@ import option from './option/index'
 
 /**
  * 创建规则
- * @param {object} option RuleData initdata
+ * @param {object} ruleOption RuleData initdata
  * @param {string} prop 规则保存属性,不存在不保存
  * @returns {RuleData}
  */
-function buildRule(option, prop) {
-  let ruleItem = new RuleData(option)
+function buildRule(ruleOption, prop) {
+  let ruleItem = new RuleData(ruleOption)
   if (prop) {
     option.data[prop] = ruleItem
   }