+++ src/data/utils-ts/isError.ts
@@ -1,6 +1,0 @@
-
-function isError(data:any): data is Error {
-  return Object.prototype.toString.call(data) === '[object Error]'
-}
-
-export default isError