+++ history.md
@@ -1,5 +1,7 @@
 ### 2.0.18ing
 - current注释
+- jsonToForm无法递归,使用JSON字符串实现逻辑
+- printMsgAct为引用isError的错误
 
 ### 2.0.17
 - 不进行测试,测试和加载在开发环境下进行